requirements.txt
aiofiles==24.1.0
python-dotenv==1.1.1
anyio==4.11.0
async-timeout==5.0.1
exceptiongroup==1.3.0